unit test refactoring verify assertions, splitting up FilteringTermsService tests
