+++ tests/Hutch.Relay.Tests/Services/FilteringTermsServiceTests/ListTests.cs
@@ -1,0 +1,86 @@
+using System.Data.Common;
+using Hutch.Relay.Config.Beacon;
+using Hutch.Relay.Data;
+using Hutch.Relay.Services;
+using Hutch.Relay.Services.Contracts;
+using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.Options;
+using Moq;
+using Xunit;
+
+namespace Hutch.Relay.Tests.Services.FilteringTermsServiceTests;
+
+public class ListTests : IDisposable
+{
+  private readonly DbConnection? _connection = null;
+
+  private static readonly IOptions<RelayBeaconOptions> DefaultOptions = Options.Create<RelayBeaconOptions>(new()
+  {
+    Enable = true
+  });
+
+  private readonly ApplicationDbContext _dbContext;
+
+  public ListTests()
+  {
+    // Ensure a unique DB per Test
+    _dbContext = FixtureHelpers.NewDbContext(ref _connection);
+    _dbContext.Database.EnsureCreated();
+  }
+
+  public void Dispose()
+  {
+    _dbContext.Database.EnsureDeleted();
+    _connection?.Dispose();
+  }
+
+  [Theory]
+  [InlineData(true)]
+  [InlineData(false)]
+  public async Task List_WhenBeaconDisabled_Warns(bool isBeaconEnabled)
+  {
+    var logger = new Mock<ILogger<FilteringTermsService>>();
+
+    RelayBeaconOptions beaconOptions = new() { Enable = isBeaconEnabled };
+
+    var service = new FilteringTermsService(
+      logger.Object,
+      Options.Create(beaconOptions),
+      Mock.Of<ISubNodeService>(),
+      Mock.Of<IDownstreamTaskService>(),
+      _dbContext
+    );
+
+    await service.List();
+
+    logger.Verify(
+      x => x.Log<It.IsAnyType>( // Must use logger.Log<It.IsAnyType> to sub-out FormattedLogValues, the internal class
+        LogLevel.Warning, // Match whichever log level you want here
+        0, // EventId
+        It.Is<It.IsAnyType>((o, t) => string.Equals(
+          "GA4GH Beacon Functionality is disabled; returning empty Filtering Terms list", o.ToString())), // The type here must match the `logger.Log<T>` type used above
+        null, //It.IsAny<Exception>(), // Whatever exception may have been logged with it, change as needed.
+        (Func<It.IsAnyType, Exception?, string>)It.IsAny<object>()), // The message formatter
+      isBeaconEnabled ? Times.Never : Times.Once);
+  }
+
+  [Fact]
+  public async Task List_WhenBeaconDisabled_ReturnsEmpty()
+  {
+    var logger = new Mock<ILogger<FilteringTermsService>>();
+
+    RelayBeaconOptions beaconOptions = new() { Enable = false };
+
+    var service = new FilteringTermsService(
+      logger.Object,
+      Options.Create(beaconOptions),
+      Mock.Of<ISubNodeService>(),
+      Mock.Of<IDownstreamTaskService>(),
+      _dbContext
+    );
+
+    var actual = await service.List();
+
+    Assert.Empty(actual);
+  }
+}
